Update GitHub Actions workflow to use staging cache for Docker images, and refactor payment and ticket actions to improve organization and maintainability.
